pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.10.0,<4.0.0"

dependencies = [
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
    "cryptography==46.0.3",
    "dataclasses-json==0.6.7",
    "faiss-cpu==1.12.0",
    "httpx-sse==0.4.3",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.0.2",
    "langchain-classic==1.0.0",
    "langchain-community==0.4",
    "langchain-core==1.0.0",
    "langchain-openai==1.0.1",
    "langchain-text-splitters==1.0.0",
    "langgraph==1.0.1",
    "langgraph-checkpoint==3.0.0",
    "langgraph-prebuilt==1.0.1",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.38",
    "marshmallow==3.26.1",
    "mypy-extensions==1.1.0",
    "orjson==3.11.3",
    "ormsgpack==1.11.0",
    "pydantic-settings==2.11.0",
    "pypdf==6.1.3",
    "requests-toolbelt==1.0.0",
    "sqlalchemy==2.0.44",
    "tenacity==9.1.2",
    "tiktoken==0.12.0",
    "typing-inspect==0.9.0",
    "xxhash==3.6.0",
    "zstandard==0.25.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
